Made binary name and config.h source configurable.

+++ Makefile
@@ -1,27 +1,27 @@
-dst = shen_run
+name = shen_run
+cfgdir = .
 
 CFLAGS = -Wall
+CFLAGS += -I$(cfgdir)
 LDFLAGS += -lutil
 destdir=/usr/local
 bindir=$(destdir)/bin
 mandir=$(destdir)/share/man
 
-all: $(dst)
+all: $(name)
 
 install:
-	install -m 755 $(dst) $(bindir)
+	install -m 755 $(name) $(bindir)
 	install -m 644 shen_run.1 $(mandir)/man1
 
 clean:
-	rm -f script.h $(dst) 2>/dev/null
+	rm -f script.h $(name) 2>/dev/null
 
-config.h:
-	cp config.def.h $@
+$(cfgdir)/config.h:
+	test -f $@ || cp config.def.h $@
 
-shen_run: shen_run.c config.h script.h
+$(name): shen_run.c $(cfgdir)/config.h script.h
+	$(CC) $(CFLAGS) $< $(LDFLAGS) -o $@
 
 script.h: script.shen
 	./mkrun <$< >$@
-
-%: %.c
-	$(CC) $(CFLAGS) $< $(LDFLAGS) -o $@
